cronograma de trabajo

+++ app/controllers/TrabajoCronogramaController.php
@@ -1,6 +1,6 @@
 <?php
 
-class ProyectoCronogramaController extends \BaseController {
+class TrabajoCronogramaController extends \BaseController {
 
 	/**
 	 * Display a listing of the resource.
@@ -26,13 +26,12 @@ class ProyectoCronogramaController extends \BaseController {
 			// Verifico si el usuario es un Webmaster
 			if($data["user"]->idrol == 1 || $data["user"]->idrol == 2 || $data["user"]->idrol == 3 || $data["user"]->idrol == 4){
 
-				$data["categorias"] = ProyectoCategoria::orderBy('nombre')->get()->lists('nombre','id');
 				$data["servicios"] = Servicio::orderBy('nombre')->get()->lists('nombre','idservicio');
 				$data["departamentos"] = Area::orderBy('nombre')->get()->lists('nombre','idarea');
 				$data["usuarios"] = User::orderBy('nombre')->get()->lists('UserFullName','id');
-				$data["proyecto"] = Proyecto::find($id);
+				$data["reporte"] = ReporteSeguimiento::find($id);
 
-				return View::make('investigacion.proyecto.cronograma.create',$data);
+				return View::make('investigacion.trabajo.create',$data);
 			}else{
 				return View::make('error/error',$data);
 			}
@@ -57,7 +56,6 @@ class ProyectoCronogramaController extends \BaseController {
 				// Validate the info, create rules for the inputs	
 				$rules = array(
 							'nombre' => 'required',
-							'categoria' => 'required',
 							'departamento' => 'required',
 							'responsable' => 'required',
 							'servicio_clinico' => 'required',
@@ -73,26 +71,108 @@ class ProyectoCronogramaController extends \BaseController {
 				$validator = Validator::make(Input::all(), $rules, $messages);
 				// If the validator fails, redirect back to the form
 				if($validator->fails()){
-					return Redirect::to('proyecto_cronograma/create/'.$id)->withErrors($validator)->withInput(Input::all());					
+					return Redirect::to('trabajo_cronograma/create/'.$id)->withErrors($validator)->withInput(Input::all());					
 				}else{
-					$proyecto_cronograma = new Cronograma;
-					$proyecto_cronograma->nombre = Input::get('nombre');
-					$proyecto_cronograma->id_categoria = Input::get('categoria');
-					$proyecto_cronograma->id_servicio_clinico = Input::get('servicio_clinico');
-					$proyecto_cronograma->id_departamento = Input::get('departamento');
-					$proyecto_cronograma->id_responsable = Input::get('responsable');
-					$proyecto_cronograma->fecha_ini = date("Y-m-d",strtotime(Input::get('fecha_ini')));
-					$proyecto_cronograma->fecha_fin = date("Y-m-d",strtotime(Input::get('fecha_fin')));
-					$proyecto_cronograma->id_proyecto = $id;
+					$trabajo_cronograma = new TrabajoCronograma;
+					$trabajo_cronograma->nombre = Input::get('nombre');
+					$trabajo_cronograma->id_servicio_clinico = Input::get('servicio_clinico');
+					$trabajo_cronograma->id_departamento = Input::get('departamento');
+					$trabajo_cronograma->id_responsable = Input::get('responsable');
+					$trabajo_cronograma->fecha_ini = date("Y-m-d",strtotime(Input::get('fecha_ini')));
+					$trabajo_cronograma->fecha_fin = date("Y-m-d",strtotime(Input::get('fecha_fin')));
+					$trabajo_cronograma->id_reporte = $id;
 
-					$proyecto_cronograma->save();
+					$trabajo_cronograma->save();
 
-					$proyecto = Proyecto::find($id);
-					$proyecto->id_cronograma = $proyecto_cronograma->id;
-					$proyecto->save();
+					$reporte = ReporteSeguimiento::find($id);
+					$reporte->id_cronograma = $trabajo_cronograma->id;
+					$reporte->save();
 					
 					Session::flash('message', 'Se registró correctamente el cronograma.');
-					return Redirect::to('proyecto_cronograma/show/'.$id);
+					return Redirect::to('trabajo_cronograma/show/'.$trabajo_cronograma->id);
+				}
+			}else{
+				return View::make('error/error',$data);
+			}
+
+		}else{
+			return View::make('error/error',$data);
+		}
+	}
+
+
+	/**
+	 * Show the form for creating a new resource.
+	 *
+	 * @return Response
+	 */
+	public function editCronograma($id)
+	{
+		if(Auth::check()){
+			$data["inside_url"] = Config::get('app.inside_url');
+			$data["user"] = Session::get('user');
+			// Verifico si el usuario es un Webmaster
+			if($data["user"]->idrol == 1 || $data["user"]->idrol == 2 || $data["user"]->idrol == 3 || $data["user"]->idrol == 4){
+
+				$data["servicios"] = Servicio::orderBy('nombre')->get()->lists('nombre','idservicio');
+				$data["departamentos"] = Area::orderBy('nombre')->get()->lists('nombre','idarea');
+				$data["usuarios"] = User::orderBy('nombre')->get()->lists('UserFullName','id');
+				$data["cronograma"] = TrabajoCronograma::find($id);
+
+				return View::make('investigacion.trabajo.editCronograma',$data);
+			}else{
+				return View::make('error/error',$data);
+			}
+		}else{
+			return View::make('error/error',$data);
+		}
+	}
+
+
+	/**
+	 * Store a newly created resource in storage.
+	 *
+	 * @return Response
+	 */
+	public function updateCronograma($id)
+	{
+		if(Auth::check()){
+			$data["inside_url"] = Config::get('app.inside_url');
+			$data["user"] = Session::get('user');
+			// Verifico si el usuario es un Webmaster
+			if($data["user"]->idrol == 1 || $data["user"]->idrol == 2 || $data["user"]->idrol == 3 || $data["user"]->idrol == 4){
+				// Validate the info, create rules for the inputs	
+				$rules = array(
+							'nombre' => 'required',
+							'departamento' => 'required',
+							'responsable' => 'required',
+							'servicio_clinico' => 'required',
+							'fecha_ini' => 'required',
+							'fecha_fin' => 'required',						
+						);
+				$messages = array(
+						'fecha_ini.required'	=> 'El campo Fecha Inicio es requerido.',
+						'fecha_fin.required'	=> 'El campo Fecha Final es requerido.',
+						
+					);
+				// Run the validation rules on the inputs from the form
+				$validator = Validator::make(Input::all(), $rules, $messages);
+				// If the validator fails, redirect back to the form
+				if($validator->fails()){
+					return Redirect::to('trabajo_cronograma/edit/'.$id)->withErrors($validator)->withInput(Input::all());					
+				}else{
+					$trabajo_cronograma = TrabajoCronograma::find($id);
+					$trabajo_cronograma->nombre = Input::get('nombre');
+					$trabajo_cronograma->id_servicio_clinico = Input::get('servicio_clinico');
+					$trabajo_cronograma->id_departamento = Input::get('departamento');
+					$trabajo_cronograma->id_responsable = Input::get('responsable');
+					$trabajo_cronograma->fecha_ini = date("Y-m-d",strtotime(Input::get('fecha_ini')));
+					$trabajo_cronograma->fecha_fin = date("Y-m-d",strtotime(Input::get('fecha_fin')));
+
+					$trabajo_cronograma->save();
+					
+					Session::flash('message', 'Se editó correctamente el cronograma.');
+					return Redirect::to('trabajo_cronograma/show/'.$trabajo_cronograma->id);
 				}
 			}else{
 				return View::make('error/error',$data);
@@ -118,15 +198,13 @@ class ProyectoCronogramaController extends \BaseController {
 			// Verifico si el usuario es un Webmaster
 			if($data["user"]->idrol == 1 || $data["user"]->idrol == 2 || $data["user"]->idrol == 3 || $data["user"]->idrol == 4){
 
-				$data["categorias"] = ProyectoCategoria::all()->lists('nombre','id');
 				$data["servicios"] = Servicio::all()->lists('nombre','idservicio');
 				$data["departamentos"] = Area::all()->lists('nombre','idarea');
 				$data["usuarios"] = User::all()->lists('UserFullName','id');
 				
-				$proyecto = Proyecto::find($id);
-				$data["cronograma"] = $proyecto->cronograma;
+				$data["cronograma"] = TrabajoCronograma::find($id);
 
-				return View::make('investigacion.proyecto.cronograma.show',$data);
+				return View::make('investigacion.trabajo.show',$data);
 			}else{
 				return View::make('error/error',$data);
 			}
@@ -151,10 +229,10 @@ class ProyectoCronogramaController extends \BaseController {
 			if($data["user"]->idrol == 1 || $data["user"]->idrol == 2 || $data["user"]->idrol == 3 || $data["user"]->idrol == 4){
 
 				$data["usuarios"] = User::orderBy('nombre')->get()->lists('UserFullName','id');
-				$data["cronograma"] = Cronograma::find($id);
-				$data["tipos"]	= [0=>'Fase de inversión',1=>'Fase de post-inversión'];
+				$data["cronograma"] = TrabajoCronograma::find($id);
+				$data["actividades"] = [0=>'No posee actividad previa'] + $data["cronograma"]->actividades->lists('nombre','id');
 
-				return View::make('investigacion.proyecto.cronograma.edit',$data);
+				return View::make('investigacion.trabajo.edit',$data);
 			}else{
 				return View::make('error/error',$data);
 			}
@@ -186,7 +264,6 @@ class ProyectoCronogramaController extends \BaseController {
 							'fecha_ini' => 'required',
 							'fecha_fin' => 'required',
 							'duracion'	=> 'required',
-							'tipo' => 'required',
 						);
 				$messages = array(
 						'fecha_ini.required'	=> 'El campo Fecha Inicio es requerido.',
@@ -196,10 +273,10 @@ class ProyectoCronogramaController extends \BaseController {
 				$validator = Validator::make(Input::all(), $rules, $messages);
 				// If the validator fails, redirect back to the form
 				if($validator->fails()){
-					return Redirect::to('proyecto_cronograma/edit/'.$id)->withErrors($validator)->withInput(Input::all());					
+					return Redirect::to('trabajo_cronograma/edit/'.$id)->withErrors($validator)->withInput(Input::all());					
 				}else{
 
-					$cronograma_actividad = new CronogramaActividad;
+					$cronograma_actividad = new TrabajoCronogramaActividad;
 
 					$cronograma_actividad->nombre = Input::get('actividad');
 					$cronograma_actividad->descripcion = Input::get('descripcion');
@@ -207,14 +284,13 @@ class ProyectoCronogramaController extends \BaseController {
 					$cronograma_actividad->fecha_ini = date("Y-m-d",strtotime(Input::get('fecha_ini')));
 					$cronograma_actividad->fecha_fin = date("Y-m-d",strtotime(Input::get('fecha_fin')));
 					$cronograma_actividad->duracion = Input::get('duracion');
-					$cronograma_actividad->id_tipo = Input::get('tipo');
 					$cronograma_actividad->id_cronograma = $id;
 
 					$cronograma_actividad->save();
 					
 					$cronograma = Cronograma::find($id);
 					Session::flash('message', 'Se editó correctamente el cronograma.');
-					return Redirect::to('proyecto_cronograma/show/'.$cronograma->proyecto->id);
+					return Redirect::to('trabajo_cronograma/show/'.$id);
 				}
 			}else{
 				return View::make('error/error',$data);
@@ -223,6 +299,18 @@ class ProyectoCronogramaController extends \BaseController {
 		}else{
 			return View::make('error/error',$data);
 		}
+	}
+
+
+	/**
+	 * Remove the specified resource from storage.
+	 *
+	 * @param  int  $id
+	 * @return Response
+	 */
+	public function destroy($id)
+	{
+		//
 	}
 
 
@@ -240,16 +328,11 @@ class ProyectoCronogramaController extends \BaseController {
 			// Verifico si el usuario es un Webmaster
 			if($data["user"]->idrol == 1 || $data["user"]->idrol == 2 || $data["user"]->idrol == 3 || $data["user"]->idrol == 4){
 
-				$data["actividad"] = CronogramaActividad::find($id);
-				$data["cronograma"] = Cronograma::find($data["actividad"]->id_cronograma);
-				if($data["actividad"]->id_tipo == 0){
-					$data["actividades"] = $data["cronograma"]->actividades->lists('nombre','id');
-				}else{
-					$data["actividades"] = $data["cronograma"]->actividadesrh->lists('nombre','id');
-				}
-				
+				$data["actividad"] = TrabajoCronogramaActividad::find($id);
+				$data["cronograma"] = TrabajoCronograma::find($data["actividad"]->id_cronograma);
+				$data["actividades"] = $data["cronograma"]->actividades->lists('nombre','id');	
 
-				return View::make('investigacion.proyecto.cronograma.editActividad',$data);
+				return View::make('investigacion.trabajo.editActividad',$data);
 			}else{
 				return View::make('error/error',$data);
 			}
@@ -290,10 +373,10 @@ class ProyectoCronogramaController extends \BaseController {
 				$validator = Validator::make(Input::all(), $rules, $messages);
 				// If the validator fails, redirect back to the form
 				if($validator->fails()){
-					return Redirect::to('proyecto_cronograma/edit/actividad/'.$id)->withErrors($validator)->withInput(Input::all());					
+					return Redirect::to('trabajo_cronograma/edit/actividad/'.$id)->withErrors($validator)->withInput(Input::all());					
 				}else{
 
-					$cronograma_actividad = CronogramaActividad::find($id);
+					$cronograma_actividad = TrabajoCronogramaActividad::find($id);
 
 					$cronograma_actividad->nombre = Input::get('actividad');
 					$cronograma_actividad->descripcion = Input::get('descripcion');
@@ -304,9 +387,9 @@ class ProyectoCronogramaController extends \BaseController {
 
 					$cronograma_actividad->save();
 					
-					$cronograma = Cronograma::find($cronograma_actividad->id_cronograma);
+					$cronograma = TrabajoCronograma::find($cronograma_actividad->id_cronograma);
 					Session::flash('message', 'Se editó correctamente la actividad del cronograma.');
-					return Redirect::to('proyecto_cronograma/show/'.$cronograma->proyecto->id);
+					return Redirect::to('trabajo_cronograma/show/'.$cronograma->id);
 				}
 			}else{
 				return View::make('error/error',$data);
@@ -315,18 +398,6 @@ class ProyectoCronogramaController extends \BaseController {
 		}else{
 			return View::make('error/error',$data);
 		}
-	}
-
-
-	/**
-	 * Remove the specified resource from storage.
-	 *
-	 * @param  int  $id
-	 * @return Response
-	 */
-	public function destroy($id)
-	{
-		//
 	}
 
 
@@ -343,11 +414,17 @@ class ProyectoCronogramaController extends \BaseController {
 			$data["user"] = Session::get('user');
 			// Verifico si el usuario es un Webmaster
 			if($data["user"]->idrol == 1 || $data["user"]->idrol == 2 || $data["user"]->idrol == 3 || $data["user"]->idrol == 4){
-				$actividad = CronogramaActividad::find($id);
-				$url = "proyecto_cronograma/show/".$actividad->cronograma->proyecto->id;
-				$actividad->delete();
-				Session::flash('message','Se borro correctamente la actividad.');					
-				return Redirect::to($url);
+				$actividad = TrabajoCronogramaActividad::find($id);
+				$url = "trabajo_cronograma/show/".$actividad->cronograma->id;
+
+				if($actividad->actividadesPosteriores->isEmpty()){
+					$actividad->delete();
+					Session::flash('message','Se borro correctamente la actividad.');					
+					return Redirect::to($url);
+				}else{
+					Session::flash('error','La actividad posee actividades posteriores.');					
+					return Redirect::to($url);
+				}
 			}else{
 				return View::make('error/error',$data);
 			}
@@ -357,7 +434,7 @@ class ProyectoCronogramaController extends \BaseController {
 	}
 
 
-	public function getActividadesAjax()
+	public function getActividadAjax()
 	{
 		
 		if(!Request::ajax() || !Auth::check())
@@ -371,27 +448,21 @@ class ProyectoCronogramaController extends \BaseController {
 		if($data["user"]->idrol == 1  || $data["user"]->idrol == 2 || $data["user"]->idrol == 3 || $data["user"]->idrol == 4  || $data["user"]->idrol == 5 || $data["user"]->idrol == 6 || $data["user"]->idrol == 7
 				 || $data["user"]->idrol == 8 || $data["user"]->idrol == 9 || $data["user"]->idrol == 10 || $data["user"]->idrol == 11 || $data["user"]->idrol == 12){
 			// Check if the current user is the "System Admin"
-			$id_cronograma = Input::get('id_cronograma');
-			$id_tipo = Input::get('id_tipo');
+			$id_actividad = Input::get('id_actividad');
 
-			if($id_cronograma!="" && $id_tipo!=""){
+			if($id_actividad!=""){
 
-				$cronograma = Cronograma::find($id_cronograma);
-				if($id_tipo == 0){
-					$actividades = [0=>'No posee actividad previa'] + $cronograma->actividades->lists('nombre','id');
-				}else{
-					$actividades = [0=>'No posee actividad previa'] + $cronograma->actividadespost->lists('nombre','id');
-				}
+				$actividad = TrabajoCronogramaActividad::find($id_actividad);
+
 				
 			}else{
-				$actividades = [0=>'No posee actividad previa'];
+				$actividad = null;
 			}
 
-			return Response::json(array( 'success' => true, 'actividades' => $actividades ),200);
+			return Response::json(array( 'success' => true, 'actividad' => $actividad ),200);
 		}else{
 			return Response::json(array( 'success' => false ),200);
 		}
 	}
-
 
 }
